requirements.txt
discord.py
pendulum
